Add new WhatsOn controller for new routes
