+++ frontend/app/controllers/Event.scala
@@ -141,55 +141,6 @@ trait Event extends Controller with ActivityTracking {
     events.sortWith(_.event.start < _.event.start)
   }
 
-  def whatsOn = GoogleAuthenticatedStaffAction { implicit request =>
-
-    val pageInfo = PageInfo(
-      CopyConfig.copyTitleEvents,
-      request.path,
-      Some(CopyConfig.copyDescriptionEvents)
-    )
-
-    val now = DateTime.now
-
-    val events = EventCollections(
-      // unstruct_event_1.eventSource:"viewEventDetails"
-      trending=guLiveEvents.getEventsByIds(List(
-        "15926171608",
-        "16351430569",
-        "15351696337",
-        "15756402825",
-        "16253355223",
-        "16234171845",
-        "16349419554",
-        "15597340064",
-        "16253236869",
-        "16430264363"
-      )),
-      // unstruct_event_1.eventSource:"eventThankYou"
-      topSelling=guLiveEvents.getEventsByIds(List(
-        "16253236869",
-        "16253494640",
-        "16234171845",
-        "16351430569",
-        "15351696337",
-        "16252865759",
-        "16253323127",
-        "15926171608",
-        "15597340064",
-        "16380673034"
-      )),
-      thisWeek=guLiveEvents.getEventsBetween(new Interval(now, now + 1.week)),
-      nextWeek=guLiveEvents.getEventsBetween(new Interval(now + 1.week, now + 2.weeks)),
-      recentlyCreated=guLiveEvents.getRecentlyCreated(now - 1.weeks),
-      partnersOnly=guLiveEvents.getEvents.filter(_.internalTicketing.exists(_.isCurrentlyAvailableToPaidMembersOnly)),
-      programmingPartnerEvents=guLiveEvents.getPartnerEvents
-    )
-
-    val latestArticles = GuardianContentService.membershipFrontContent.map(ContentItem)
-
-    Ok(views.html.event.whatson(pageInfo, events, latestArticles))
-  }
-
   def list = CachedAction { implicit request =>
     val pageInfo = PageInfo(
       CopyConfig.copyTitleEvents,
